backend/requirements.txt
flask==3.1.0
flask-cors==4.0.1
numpy>=2.0.0
opencv-python-headless>=4.10.0.84
Pillow>=10.4.0
